// src/wmslang/builtin_catalog.generated.ts
export type WmslangBuiltinValueType = "f32" | "f32x2" | "f32x3" | "f32x4" | "i32" | "i32x2" | "i32x3" | "i32x4";